tests/wpt/web-platform-tests/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/img-no-alt-replaced.html

<!doctype html>
<title>Images without alt attribute or with an empty alt attribute render as replaced elements regardless of src</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1196668">
<style>
  img {
    width: 100px;
    height: 150px;
  }
</style>
<img>
<img src="broken">
<img alt="">
<script>
const t = async_test("Images without alt attribute or with an empty alt attribute render as replaced elements regardless of src");
onload = t.step_func_done(function() {
  for (const img of document.querySelectorAll("img")) {
    assert_equals(img.offsetWidth, 100, `width: ${img.outerHTML}`);
    assert_equals(img.offsetHeight, 150, `height: ${img.outerHTML}`);
  }
});
</script>
