LayoutTests/imported/w3c/web-platform-tests/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/img-no-alt-replaced-expected.txt

    

FAIL Images without alt attribute or with an empty alt attribute render as replaced elements regardless of src assert_equals: width: <img src="broken" alt=""> expected 100 but got 0
